removed the different sections
